Rename `PlayerUtils` to `NumPadUtils` to reflect its functionality better
